options/ansi: move strptime to options/posix
